+++ context.py
@@ -1,0 +1,3 @@
+class AppContext:
+    inventory = None
+    master_data = None
