+++ master.py
@@ -1,8 +1,8 @@
 import os
-import helpers
+import helpers.utils as utils
 import csv
 
-MASTER_DATA_PATH = "data/master"
+MASTER_DATA_PATH = "../data/master"  # Rel. to parse_file() func
 
 class MasterData:
     def __init__(self):
@@ -10,7 +10,18 @@ class MasterData:
         self.set_items = init_master_set_items()
         self.unique_items = init_master_unique_items()
         # self.other_items = helpers.parse_file(f"{MASTER_DATA_PATH}/other.txt")
-        
+
+def init_all_set_names_in_collection(set_items_raw, set_items):
+    """index	*ID	set	item	*ItemName	rarity
+    First few entries in the "labels" rowin master data. We are mainly 
+    interested in the "set" entry at index 2 for each row here.
+    """
+    all_sets = set()
+    for item in set_items_raw:
+        entries = item.split('\t')
+        if (len(entries) > 2 and entries[2] != ''):
+            all_sets.add(entries[2])
+
 def init_master_set_items():
     """Initialize the master set items dictionary from raw data.
 
@@ -18,9 +29,11 @@ def init_master_set_items():
     a dictionary where each item is keyed by its name. Missing or empty fields 
     in the raw data are handled by assigning "N/A" as a default value.
     """
-    set_items_raw = helpers.parse_file(f"{MASTER_DATA_PATH}/set.txt")
+    set_items_raw = utils.parse_file(f"{MASTER_DATA_PATH}/set.txt")
+    labels = set_items_raw.pop(0).split('\t')  # Pop the row with labels, careful if moving this
+    
     set_items = {}
-    labels = set_items_raw.pop(0).split('\t')
+    set_items = init_all_set_names_in_collection(set_items_raw, set_items)
     
     for idx, item in enumerate(set_items_raw):
         entries = item.split('\t')
@@ -36,7 +49,7 @@ def init_master_set_items():
 
 def init_master_unique_items():
     """Initialize master unique items, same handling as for set items"""
-    unique_items_raw = helpers.parse_file(f"{MASTER_DATA_PATH}/unique.txt")
+    unique_items_raw = utils.parse_file(f"{MASTER_DATA_PATH}/unique.txt")
     unique_items = {}
     labels = unique_items_raw.pop(0).split('\t')
